Fix narrowing conversion warning

+++ midi-octopot/midi-octopot.ino
@@ -27,7 +27,7 @@ byte pot_val[] = {0, 0, 0, 0, 0, 0, 0, 0};
 
 void controlChange(byte channel, byte control, byte value) {
 
-  midiEventPacket_t event = {0x0B, 0xB0 | channel, control, value};
+  midiEventPacket_t event = {0x0B, (byte)(0xB0 | channel), control, value};
 
   MidiUSB.sendMIDI(event);
   MidiUSB.flush();
